feat : Add record topic. Fix bug of point type

+++ src/planner/local_planner/src/pathFollower.cpp
@@ -126,14 +126,14 @@ ros::Time sendTime;
 
 nav_msgs::Path path;
 
-std::vector<std::vector<cv::Point2d>> polygons =
+std::vector<std::vector<cv::Point2f>> polygons =
 {
   // 3楼梯高
   {
-    cv::Point2d(-3.4, 1.50),
-    cv::Point2d(-3.24, 0.42),
-    cv::Point2d(-1.65, 0.60),
-    cv::Point2d(-1.00, 1.51)
+    cv::Point2f(-3.4, 1.50),
+    cv::Point2f(-3.24, 0.42),
+    cv::Point2f(-1.65, 0.60),
+    cv::Point2f(-1.00, 1.51)
   },
 };
